Solved: Exercises 4.15-4.21 (Bloglist app backend) - Implemented user administration and user login through token   authentication - Implemented validation when adding new user:   > Both username and password must be given   > Both username and password must be minimum 3 characters   > Username must be unique - Expanded test suite to verify above validation - Expanded blog and user route handlers as follows:   > Blog can only be created by logged-in user via valid token   > When blog is created, add logged-in user as creator   > Deleting blog only allowed by creator user via valid token   > Listing blogs will show creator's user info within the blog   > Listing users will show blogs created by each user - Refactored getTokenFrom() helper function to middleware

+++ part4/bloglist-backend/controllers/blogs.js
@@ -1,9 +1,12 @@
+const jwt = require('jsonwebtoken');
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
+const User = require('../models/user');
 
 // get all blogs
 blogsRouter.get('/', async (req, res) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 });
   res.json(blogs);
 });
 
@@ -20,6 +23,15 @@ blogsRouter.get('/:id', async (req, res) => {
 // add new blog
 blogsRouter.post('/', async (req, res) => {
   const body = req.body;
+  // eslint-disable-next-line no-undef
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({
+      error: 'token missing or invalid'
+    });
+  }
+
+  const user = await User.findById(decodedToken.id);
 
   if (!body.title || !body.author || !body.url) {
     return res.status(400).json({
@@ -31,16 +43,39 @@ blogsRouter.post('/', async (req, res) => {
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    user: user._id
   });
 
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
   res.json(savedBlog);
 });
 
 // delete blog
 blogsRouter.delete('/:id', async (req, res) => {
+  // eslint-disable-next-line no-undef
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({
+      error: 'token missing or invalid'
+    });
+  }
+
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(req.params.id);
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return res.status(401).json({
+      error: 'blog can be deleted only by user who created it'
+    });
+  }
+
   await Blog.findByIdAndRemove(req.params.id);
+  user.blogs = user.blogs.filter(b => b._id.toString() !== req.params.id);
+  await user.save();
   res.status(204).end();
 });
